face-recognition/requirements.txt
click==7.1.1
dlib==19.19.0
face-recognition==1.3.0
face-recognition-models==0.3.0
numpy==1.18.2
Pillow==7.0.0
pkg-resources==0.0.0
